requirement.txt
tensorboard
easydict
sklearn
matplotlib
pandas
scikit-image
menpo
prettytable
mxnet==1.6.0
pytorch_summary
cv2